merge

+++ trident/data/utils.py
@@ -242,10 +242,11 @@ def download_model_from_google_drive(file_id, dirname, filename=None, md5=None):
     try:
         session = requests.Session()
         response = session.get(url, params={'id': '12XLjt9Zcaoo90WGG6R5N0U6Sf_KBZZn_'}, stream=False)
-        if os.path.exists(os.path.join(dirname, 'models_md5.json')):
-            os.remove(os.path.join(dirname, 'models_md5.json'))
-        with open(os.path.join(dirname, 'models_md5.json'),"wb") as f:
-            f.write(response.content)
+        if response.status_code==200:
+            if os.path.exists(os.path.join(dirname, 'models_md5.json')):
+                os.remove(os.path.join(dirname, 'models_md5.json'))
+            with open(os.path.join(dirname, 'models_md5.json'),"wb") as f:
+                f.write(response.content)
 
         check_internet = True
         if os.path.exists(os.path.join(dirname, 'models_md5.json')):
